+++ R/hashing_dummy.R
@@ -189,6 +189,7 @@ bake.step_dummy_hash <- function(object, new_data, ...) {
         object$num_terms
       )
 
+    tf_text <- purrr::map_dfc(tf_text, as.integer)
     keep_original_cols <- get_keep_original_cols(object)
     if (!keep_original_cols) {
       new_data <-